refactor(backend): consolidate provider cloud/local classification into single source of truth
- Delete hardcoded CLOUD_PROVIDER_IDS and LOCAL_PROVIDER_IDS sets from model-router.mjs
- Export single isCloudProviderId helper from base.mjs
- Delegate AIService.isCloudProvider and model-router functions to isCloudProviderId
