+++ node-backend/lib/ai.mjs
@@ -1,3 +1,4 @@
+import { isCloudProviderId } from './providers/base.mjs';
 import { OllamaProvider } from './providers/ollama.mjs';
 import { HostedAPIProvider } from './providers/hosted.mjs';
 import { GoogleCliProvider } from './providers/google.mjs';
@@ -47,26 +48,7 @@ export class AIService {
   }
 
   static isCloudProvider(providerType, settings = {}) {
-    const type = String(providerType || '');
-    if (type === 'ollama') return false;
-
-    const customClis = settings.customClis || settings.custom_clis || [];
-    if (Array.isArray(customClis)) {
-      const custom = customClis.find(c =>
-        c.id === type ||
-        `custom-${c.id}` === type ||
-        c.name === type ||
-        `custom-${c.name}` === type
-      );
-      if (custom) {
-        if (typeof custom.isCloud === 'boolean') return custom.isCloud;
-        if (type.startsWith('custom-local-') || custom.id?.startsWith('custom-local-') || custom.id?.includes('local')) {
-          return false;
-        }
-      }
-    }
-    if (type.startsWith('custom-local-')) return false;
-    return true;
+    return isCloudProviderId(providerType, settings);
   }
 
   static async createProvider(providerType, settings = {}, secrets = {}) {
